have AddCustomCourse component seperately
